requirements.txt
python-telegram-bot==21.4
aiohttp==3.10.8
python-dotenv==1.0.1
pydantic==2.9.2
